add Credential model and rename existing credential
