+++ src/models/Credential.ts
@@ -1,0 +1,8 @@
+export interface CredentialEntry {
+  id: number;
+  website: string;
+  title: string;
+  username: string;
+  password: string;
+  notes?: string;
+}
